Move intervals to filter.ts
